Add player-related functionality to the engine
