--- src/parser/testmap.cub
@@ -1,23 +1,0 @@
-
-NO a.txt
-
-EA  b.txt
-
-SO c.txt
-
-C 53,  41,34
-
-F   45, 33, 27
-
-WE d.txt
-
-111111111111111111
-110000000000  000011
-110000000000000011
-110000000000000011
-1100N00000000 00011
-110000000000000011
-110000000000000011
-11000000000  0000011
-1111111111 11111111
-111111111111111111
